Split PokedexViewModel.kt into two
