+++ feature-pokedex/src/main/java/com/victor/feature_pokedex/presentation/PokedexActivity.kt
@@ -7,6 +7,8 @@ import androidx.activity.enableEdgeToEdge
 import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.runtime.Composable
 import androidx.navigation.compose.rememberNavController
+import com.victor.feature_pokedex.presentation.ui.details.DetailsViewModel
+import com.victor.feature_pokedex.presentation.ui.home.HomeViewModel
 import com.victor.feature_pokedex.presentation.ui.navigation.PokedexNavHost
 import com.victor.feature_pokedex.presentation.ui.theme.PokedexTheme
 import com.victor.features_common.viewModel
@@ -17,25 +19,27 @@ import org.kodein.di.android.closestDI
 class PokedexActivity : ComponentActivity(), DIAware {
 
     override val di by closestDI()
-    private val viewModel: PokedexViewModel by viewModel()
+    private val homeViewModel: HomeViewModel by viewModel()
+    private val detailsViewModel: DetailsViewModel by viewModel()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
         setContent {
-            PokedexApp(viewModel)
+            PokedexApp(homeViewModel, detailsViewModel)
         }
     }
 }
 
 @ExperimentalFoundationApi
 @Composable
-private fun PokedexApp(viewModel: PokedexViewModel) {
+private fun PokedexApp(homeViewModel: HomeViewModel, detailsViewModel: DetailsViewModel) {
     PokedexTheme {
         val navController = rememberNavController()
         PokedexNavHost(
             navController = navController,
-            viewModel = viewModel,
+            homeViewModel = homeViewModel,
+            detailsViewModel = detailsViewModel,
         )
     }
 }
