+++ feature-pokedex/src/main/java/com/victor/feature_pokedex/presentation/ui/details/DetailsViewModel.kt
@@ -2,22 +2,23 @@ package com.victor.feature_pokedex.presentation.ui.details
 
 import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.ViewModel
+import com.victor.feature_pokedex.domain.PokemonInformationUseCase
 import com.victor.feature_pokedex.domain.PokemonListUseCase
 import com.victor.feature_pokedex.domain.model.Pokemon
 import com.victor.features_common.State
+import com.victor.features_common.manageStateDuringRequest
 
 internal class DetailsViewModel(
-    private val details: Pokemon,
-    private val useCase: PokemonListUseCase,
+    private val pokemonInformationUseCase: PokemonInformationUseCase,
 ) : ViewModel() {
 
-    private val pokemonSpecies = mutableStateOf<State>(State.Empty)
+    val pokemonInformation = mutableStateOf<State>(State.Empty)
 
-//    fun getPokemonInformation() {
-//        manageStateDuringRequest(
-//            mutableState = pokemonSpecies,
-//        ) {
-//            useCase.getPokemonInformation(details.id)
-//        }
-//    }
+    fun getPokemonInformation(pokemonId: Int) {
+        manageStateDuringRequest(
+            mutableState = pokemonInformation,
+        ) {
+            pokemonInformationUseCase.getPokemonInformation(pokemonId)
+        }
+    }
 }
