+++ feature-pokedex/src/main/java/com/victor/feature_pokedex/presentation/ui/home/bottomsheets/FilterBottomSheet.kt
@@ -31,7 +31,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.victor.feature_pokedex.R
 import com.victor.feature_pokedex.domain.model.TypeSimple
-import com.victor.feature_pokedex.presentation.PokedexViewModel
+import com.victor.feature_pokedex.presentation.ui.home.HomeViewModel
 import com.victor.feature_pokedex.presentation.ui.components.PokemonTypeIcon
 import com.victor.features_common.components.PokedexButton
 import com.victor.features_common.components.PokedexButtonStyle
@@ -42,7 +42,7 @@ import com.victor.features_common.theme.Red
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-internal fun FilterBottomSheet(viewModel: PokedexViewModel) {
+internal fun FilterBottomSheet(viewModel: HomeViewModel) {
     ModalBottomSheet(
         onDismissRequest = { viewModel.onDismissBottomSheet() },
         sheetState = rememberModalBottomSheetState(skipPartiallyExpanded = true),
