+++ feature-pokedex/src/main/java/com/victor/feature_pokedex/presentation/ui/navigation/PokedexNavHost.kt
@@ -8,15 +8,17 @@ import androidx.navigation.NavType
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.navArgument
-import com.victor.feature_pokedex.presentation.PokedexViewModel
+import com.victor.feature_pokedex.presentation.ui.home.HomeViewModel
 import com.victor.feature_pokedex.presentation.ui.details.DetailsScreenBody
+import com.victor.feature_pokedex.presentation.ui.details.DetailsViewModel
 import com.victor.feature_pokedex.presentation.ui.home.HomeScreenBody
 
 @ExperimentalFoundationApi
 @Composable
 internal fun PokedexNavHost(
     navController: NavHostController,
-    viewModel: PokedexViewModel,
+    homeViewModel: HomeViewModel,
+    detailsViewModel: DetailsViewModel,
     modifier: Modifier = Modifier
 ) {
     NavHost(
@@ -25,7 +27,7 @@ internal fun PokedexNavHost(
         modifier = modifier
     ) {
         composable(Screens.HomeScreen.name) {
-            HomeScreenBody(viewModel = viewModel, onPokemonClick = {
+            HomeScreenBody(viewModel = homeViewModel, onPokemonClick = {
                 val route = "${Screens.DetailsScreen.name}/$it"
                 navController.navigate(route)
             })
@@ -39,7 +41,7 @@ internal fun PokedexNavHost(
             val pokemonId = it.arguments?.getInt("pokemonId") ?: 0
             DetailsScreenBody(
                 navController = navController,
-                viewModel = viewModel,
+                viewModel = detailsViewModel,
                 pokemonId = pokemonId,
                 onPokemonClick = { id ->
                     if (id != pokemonId) {
